add register function when adding member/admin
Add password field to all entry DTOs for person creation. Add JWTUserDetailsService attribute to services in charge of person creation. When creating a person, a call to JWTUserDetailsService will be made to save that person as a user of the application, using email as id, password as password and setting role as either 'familyAdministrator' or 'familyMember'. Add transactional tag where needed to make sure unit of work is used. Fix tests with new arguments as needed. Add tests for DTOs comparing password attribute.